[genio] fix log format

+++ src/platform/mt793x/SystemTimeSupport.cpp
@@ -104,10 +104,11 @@ CHIP_ERROR ClockImpl::SetClock_RealTime(Clock::Microseconds64 aNewCurTime)
         const time_t timep = tv.tv_sec;
         struct tm calendar;
         localtime_r(&timep, &calendar);
-        ChipLogProgress(
-            DeviceLayer,
-            "Real time clock set to %ld (%04" PRId16 "/%02" PRId8 "/%02" PRId8 " %02" PRId8 ":%02" PRId8 ":%02" PRId8 " UTC)",
-            static_cast<long int>(tv.tv_sec), calendar.tm_year, calendar.tm_mon, calendar.tm_mday, calendar.tm_hour, calendar.tm_min, calendar.tm_sec);
+        ChipLogProgress(DeviceLayer,
+                        "Real time clock set to %ld (%04d/%02d/%02d %02d:%02d:%02d UTC)",
+                        tv.tv_sec, calendar.tm_year,
+                        calendar.tm_mon, calendar.tm_mday,
+                        calendar.tm_hour, calendar.tm_min, calendar.tm_sec);
     }
 #endif // CHIP_PROGRESS_LOGGING
     return CHIP_NO_ERROR;
